Break circular dependency between memory and runner
Move SchemaContext and SchemaPathSelector type definitions from runner
to memory/interface.ts, eliminating the circular import between packages.

Changes:
- memory/interface.ts: Define SchemaContext and SchemaPathSelector locally,
  import JSONSchema from @commontools/api instead of @commontools/runner
- runner/src/builder/types.ts: Import SchemaContext from memory
- runner/src/traverse.ts: Import SchemaContext and SchemaPathSelector from memory

Impact:
- Files without complex runner dependencies now compile 30-40x faster
  (0.013s vs 0.5s)
- Simple utility files reduced from 267 dependencies (34MB) to 2-5
  dependencies (~116KB)
- Total runner package typecheck time reduced by 3 seconds
